# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca Virtual.",
        "description": "Essa API tem como objetivo o gerenciamento de livros em uma biblioteca virutal.",
        "contact": {
            "email": "peusodati@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3333/v1",
            "description": "API de Desenvolvimento"
        },
        {
            "url": "http://124.55.65.22:3333/v1",
            "description": "API de Produção"
        }
    ],
    "paths": {
...
        "/livros/": {
            "post": {
                "sumary": "Cadastro de um livro.",
                "description": "Esta rota é responsável por realizar o cadastro de um livro.",
                "tags": ["Livros"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schema/Livro"
                            },
                            "examples": {
                                "livro": {
                                    "value": {
                                        "coordenadorId": "8a1dfb34-c487-404d-ae7b-8464da1679f0",
                                        "nome": "O Poder do Hábito",
                                        "descricao": "O poder do hábito trás grandes reflexões sobre a vida.",
                                        "autor": "Jhon Brown"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schema/Livro"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro de validação": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "code": "tipo invalido",
                                            "expected": "string",
                                            "received": "undefined",
                                            "message": "required || O campo nome é obrigatório",
                                            "path": [
                                                "coordenadorId"
                                            ]
                                        }
                                    },
                                    "Erro de validação dados requisição": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "message": "Não existe coordenador cadastro com esse Id."
                                        }
                                    }
                                }
                            }
                        } 
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }, 
            "get": {
                "sumary": "Busca de todos os livros cadastrados.",
                "description": "Esta rota é responsável por buscar todos os livros cadastrados.",
                "tags": ["Livros"],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schema/Livro"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/livros/{id}": {
            "get": {
                "sumary": "Busca de livro por Id",
                "description": "Esta rota é responsável por buscar um único livro por Id.",
                "tags": ["Livros"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id para buscar o livro.",
                        "required": true,
                        "type": "string",
                        "format": "uuid",
                        "example": "8a1dfb34-c487-404d-ae7b-8464da1679f0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schema/Livro"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro de validação": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "validation": "uuid",
                                            "code": "invalid_string",
                                            "message": "Invalid uuid || O campo id deve ser do tipo string uuid",
                                            "path": [
                                                "id"
                                            ]
                                        }
                                    },
                                    "Erro de validação dados requisição": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "message": "Não existe livro cadastro com esse Id."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "sumary": "Atualização de livro por Id",
                "description": "Esta rota é responsável por atualizar um livro por Id.",
                "tags": ["Livros"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schema/Livro"
                            },
                            "examples": {
                                "livro": {
                                    "value": {
                                        "nome": "Dom casmurro",
                                        "autor": "Charles do Broxn", 
                                        "descricao": "Dom casmurro retrata a história do lutador de MMA Charles do Bronx"
                                    }
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id para atualizar o livro",
                        "required": true,
                        "type": "string",
                        "format": "uuid",
                        "example": "8a1dfb34-c487-404d-ae7b-8464da1679f0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schema/Livro"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação ou outro erro do cliente.",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro de validação Id": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "validation": "uuid",
                                            "code": "invalid_string",
                                            "message": "Invalid uuid || O campo id deve ser do tipo string uuid",
                                            "path": [
                                                "id"
                                            ]
                                        }
                                    },
                                    "Erro de validação": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "code": "tipo invalido",
                                            "expected": "string",
                                            "received": "number",
                                            "path": [
                                                "nome"
                                            ],
                                            "message": "Esperado string, recebido number || O campo nome é obrigatório"
                                        }
                                    },
                                    "Erro de validação dados requisição": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "message": "Não existe livro cadastro com esse Id."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "sumary": "Exclusão de um livro por Id",
                "description": "Esta rota é responsável por excluir um único livro por Id.",
                "tags": ["Livros"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id para exclusão de um coordenador",
                        "required": true,
                        "type": "string",
                        "format": "uuid",
                        "example": "8a1dfb34-c487-404d-ae7b-8464da1679f0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "message": "Livro excluído."
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro de validação": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "validation": "uuid",
                                            "code": "invalid_string",
                                            "message": "Invalid uuid || O campo id deve ser do tipo string uuid",
                                            "path": [
                                                "id"
                                            ]
                                        }
                                    },
                                    "Erro de validação dados requisição": {
                                        "sumary": "Erro na validação dos dados enviados",
                                        "value": {
                                            "message": "Não existe livro cadastro com esse Id."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Erro no Servidor": {
                                        "sumary": "Internal Server Error",
                                        "value": {
                                            "message": "Internal Server Error."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schema": {
...
            "Livro": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "autor": {
                        "type": "string"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "disponivel": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}